requirements.txt
PyQt6>=6.4.0
woocommerce>=3.0.0
Pillow>=10.0.0
requests>=2.31.0
pyobjc-framework-Cocoa>=9.0; sys_platform == "darwin"
